bug fix related to scanning
set cmd variable value July 2025- Chetanya

This has some function. DO NOT DISABLE

If you try to scan without this, the values would not be sent
to the device you are scanning as this updates the 'value' variable
after which that value is sent to the device. So if the device is stuck
at 0 and you scan from 1 to 3, it will send 0 to the device for earch step